# perf: portable_simd 3bpp paeth filter