With `--features=unstable`, this implementation will decode up to 16 pixels at once.

Reduces the number of cycles required to decode a 768px row by 20-30% on Arm systems, and around 50-70% on x64 systems ([see discussion](https://github.com/image-rs/image-png/discussions/416)

Results on an AMD EPYC 7B13 system
    $  taskset -c 64-68 cargo +nightly bench --features=benchmarks,unstable -- filter=Paeth/bpp=3
    unfilter/filter=Paeth/bpp=3
        time:   [18.987 µs 18.994 µs 19.001 µs]
        thrpt:  [616.74 MiB/s 616.96 MiB/s 617.19 MiB/s]

    $ taskset -c 64-68 cargo +nightly bench --features=benchmarks,unstable -- filter=Paeth/bpp=3
        unfilter/filter=Paeth/bpp=3
            time:   [15.355 µs 15.361 µs 15.367 µs]
            thrpt:  [762.59 MiB/s 762.89 MiB/s 763.18 MiB/s]
        change:
            time:   [-19.180% -19.087% -19.007%] (p = 0.00 < 0.05)
            thrpt:  [+23.468% +23.589% +23.732%]