@Shnatsel 2025-08-29
That is correct, `portable_simd` does not do runtime feature selection. 

It can be added as a layer on top of `portable_simd` but that will require some `unsafe` in the implementation. The [`multiversion`](https://crates.io/crates/multiversion) crate provides a safe wrapper around that using a proc macro.

I suggest exploring multiversioning separately from this PR, since it is a mostly orthogonal concern.