@Shnatsel 2025-08-30
We never really got 3bpp Paeth to vectorize, so it makes sense that large gains are possible here.

On a desktop Zen4 the difference is under 5%, but at least it seems to be stable and not noise. "My CPU is too good for optimization work" is a weird problem to have.