@Shnatsel 2025-08-30
Bad news: this is a 10% regression on an Apple M4 laptop.

Measured with `cargo +nightly bench --features=benchmarks,unstable -- filter=Paeth/bpp=3`