@Sentimentron 2025-08-30
Yes, looks like I get a similar result on my personal M2 laptop:

```
unfilter/filter=Paeth/bpp=3
                        time:   [19.663 µs 19.667 µs 19.671 µs]
                        thrpt:  [595.75 MiB/s 595.87 MiB/s 595.98 MiB/s]
                 change:
                        time:   [+11.301% +11.404% +11.494%] (p = 0.00 < 0.05)
                        thrpt:  [-10.309% -10.236% -10.154%]
                        Performance has regressed.
``` 

I don't know if the LLVM cost model for the M1 is in any way accurate, but I think we can gate this off for Mac OS for now.