@Sentimentron 2025-08-30
Looking at the public LLVM cost model, seems like all the Apple cores are modelled using this SchedCyclone def:

https://github.com/llvm/llvm-project/blame/main/llvm/lib/Target/AArch64/AArch64SchedCyclone.td

Doesn’t seem to have received any significant updates in the last 11 years or so. It’s possible that they do this properly in their proprietary Clang, but I don’t think the LLVM MCA-based simulator that I can run will accurately model this processor. This work is mostly about lifting up the Arm low-end (Cortex A520/A35/A53) which are all relatively simple to model - any extra things that benefit are just a bonus!