@Shnatsel 2025-08-30
I wonder if this is at all beneficial in practice on lower-end ARM, since older SoCs are heavily bottlenecked by memory bandwidth, so microarchitectural optimizations might be entirely irrelevant.

It's a hassle to get benchmarks from Android, but maybe we could get this benchmarked on a Raspberry Pi or some such?