@awxkee 2025-08-30
Making effective intrinsics on NEON for such a small series like `[u8; 3]` is incredible complicated even with direct intrinsics, I almost never had success with this. NEON instructions are power hungry so instead of cheap scalar instructions on big cores they consule roughly 2 times more energy to execute and cause more heat, to process such a small series. As a result, this can lead to throttling.

> Does PNG decoding get scheduled onto little cores often? I would imagine that interactive workloads like web browsing would run mostly on the big cores, with the little ones mostly getting non-interactive background processes.

It starts on the little cores, and if the workload does not decrease, it switches to the big cores (if not pinned). After that, it may throttle.