@Sentimentron 2025-08-31
> ~I'm curious what the microarchitecture simulator has to say about #635 which significantly improves benchmarks for x86_64, Apple M3 and gives an especially huge boost to Graviton 3.~ I misread the benchmarks, it actually regresses ARM a lot. Never mind.

I was curious too. Testing #635, the micro-architecture simulator indicates this:
Filter | Arm Cortex A520 | Arm Cortex X4 | Intel Arrow Lake | AMD Zen 5 | Arm Cortex A520 (baseline cycles) | Arm Cortex X4 (baseline cycles) | Intel Arrow Lake (baseline cycles) | AMD Zen 5 (baseline cycles)
-- | -- | -- | -- | -- | -- | -- | -- | --
Paeth (3bpp) | -13.92% | -18.52% | 1.35% | 0.20% | 1073543 | 520265 | 1158373 | 505037
Paeth (4bpp) | 9.53% | -7.49% | 6.28% | -4.25% | 885251 | 288701 | 413955 | 259222

The PR doesn't say which CPU it was measured on, but the results indicate a 6% uplift on both bpp=3 and bpp=4. I've measured it on my Epyc system and it yields about a 2% uplift. It's not surprising that the micro-arch simulation is a little off for the reasons I've discussed (it assumes basically a "perfect" path to main memory with no latency), this is what LLVM's cost-model _thinks_ the code will cost to execute. A theoretically worse solution can be better in practise as long as it decreases stalled cycles.