@awxkee 2025-08-31
FWIW 
To make this PR effective on NEON, it’s necessary to figure out a way to utilize the pipeline and instruction scheduling, to compensate introduced latency. On AArch64, you can’t just generate nearly the same amount of code with more power hungry instructions and expect guaranteed performance improvements. Some gains might appear sometimes, but they often come at the cost of shifting work from one bottleneck to another.
For example, instead of processing `pixel(row, 0), pixel(row, 1), pixel(row, 2)`, process `process_pixels(row0, row1, row2, row3, 0), process_pixels(row0, row1, row2, row3, 1)` to give LLVM chance to schedule instructions on the NEON pipeline, since multiple rows will be processed for the same latency ( as they are not blocking each other ) if there is more than one SIMD unit on a chip.