@Sentimentron 2025-09-01
Although, @awxkee's comment does have some truth to it. Switching off SLP auto-vectorization entirely (i.e. no neon generated):

```
$ RUSTFLAGS="-C llvm-args=--slp-threshold=30000" cargo +nightly bench --features=benchmarks -- filter=Paeth/bpp=3
unfilter/filter=Paeth/bpp=3
                        time:   [14.587 µs 14.619 µs 14.675 µs]
                        thrpt:  [798.57 MiB/s 801.59 MiB/s 803.35 MiB/s]
                 change:
                        time:   [-17.228% -17.138% -16.971%] (p = 0.00 < 0.05)
                        thrpt:  [+20.440% +20.683% +20.813%]
                        Performance has improved.
```

This indicates that... yes, someone should definitely try and model Apple Silicon a bit better in upstream LLVM: it's a very unusual CPU.